extract config loading into a function
